Migrando PrivacyStatement a frontend/src
Este cambio mueve \OmegaUp\PrivacyStatement al directorio que puede ser
autoloaded.
