+++ frontend/server/controllers/ContestController.php
@@ -1,6 +1,5 @@
 <?php
 
-require_once 'libs/PrivacyStatement.php';
 require_once 'libs/dao/Contests.dao.php';
 
 /**
@@ -376,7 +375,7 @@ class ContestController extends Controller {
         );
 
         // Privacy Statement Information
-        $privacyStatementMarkdown = PrivacyStatement::getForProblemset(
+        $privacyStatementMarkdown = \OmegaUp\PrivacyStatement::getForProblemset(
             $identity->language_id,
             'contest',
             $result['requestsUserInformation']
