+++ frontend/server/controllers/CourseController.php
@@ -1,7 +1,5 @@
 <?php
 
-require_once 'libs/PrivacyStatement.php';
-
 /**
  *  CourseController
  *
@@ -1432,7 +1430,7 @@ class CourseController extends Controller {
                 || !is_null($r->identity->state_id) || !is_null($r->identity->school_id));
 
             // Privacy Statement Information
-            $privacyStatementMarkdown = PrivacyStatement::getForProblemset(
+            $privacyStatementMarkdown = \OmegaUp\PrivacyStatement::getForProblemset(
                 $r->identity->language_id,
                 'course',
                 $requestUserInformation
@@ -1452,7 +1450,7 @@ class CourseController extends Controller {
                 $privacyStatement['statementType'] = $statementType;
             }
 
-            $markdown = PrivacyStatement::getForConsent(
+            $markdown = \OmegaUp\PrivacyStatement::getForConsent(
                 $r->identity->language_id,
                 'accept_teacher'
             );
